platform/user_core/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
arrow==1.3.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
email-validator==2.1.0
fastapi==0.119.0
fqdn==1.5.1
graphql-core==3.2.6
h11==0.16.0
harfile==0.4.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.140.3
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.11
iniconfig==2.1.0
isoduration==20.11.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
junit-xml==1.9
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.27.2
Pygments==2.19.2
pyrate-limiter==3.9.0
pytest==8.4.2
pytest-subtests==0.14.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==14.2.0
rpds-py==0.27.1
schemathesis==4.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
starlette-testclient==0.4.1
tenacity==9.0.0
tomli==2.3.0
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.37.0
webcolors==24.11.1
Werkzeug==3.1.3
